% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CarefulOptim}
\alias{CarefulOptim}
\title{Split function into k bits and optimize on each bit in case not unimodal}
\usage{
CarefulOptim(f, interval, k = 10)
}
\description{
Split function into k bits and optimize on each bit in case not unimodal
}
\keyword{internal}

