% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{cghs}
\alias{cghs}
\title{Oreopoulos (2006) UK general household survey dataset}
\format{
A data frame with 73,954 rows and 2 variables:
\describe{
  \item{earnings}{Annual earnings in 1998 (UK pounds)}
  \item{yearat14}{Year individual turned 14}
 }
}
\source{
American Economic Review data archive,
\doi{10.1257/000282806776157641}
}
\usage{
cghs
}
\description{
Oreopoulos (2006) UK general household survey dataset
}
\references{
{

\cite{Oreopoulos, P. (2006): "Estimating Average and Local Average Treatment
      Effects When Compulsory Education Schooling Laws Really Matter",
      American Economic Review, 96 (1), 152-175}

}
}
\keyword{datasets}
