% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnm.R
\name{CVb}
\alias{CVb}
\title{Critical values for CIs based on a biased Gaussian estimator.}
\usage{
CVb(B, alpha = 0.05)
}
\arguments{
\item{B}{Maximum bias, vector of non-negative numbers.}

\item{alpha}{Determines CI level, \code{1-alpha}. Needs to be between 0 and
1. Can be a vector of values.}
}
\value{
Data frame with the following columns:

\describe{

\item{bias}{Value of bias as specified by \code{bs}}

\item{alpha}{Value of \eqn{\alpha} as specified by \code{alpha}}

\item{cv}{Critical value}

\item{TeXDescription}{LaTeX-friendly description of current row}}
}
\description{
Computes the critical value \eqn{cv_{1-alpha}(B)} such that the confidence
interval \eqn{X\pm cv_{1-alpha}(B)} will have coverage \eqn{1-alpha}, where
\eqn{X} is Normally distributed with variance equal to \eqn{1} and maximum
bias at most \eqn{B}.
}
\examples{
# 90\% critical value:
CVb(B = 1, alpha = 0.1)
CVb(B = c(0, 0.5, 1), alpha = c(0.05, 0.1))
}
