% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{EqKern}
\alias{EqKern}
\title{Equivalent kernel for local linear regression.}
\usage{
EqKern(kernel = "uniform", boundary = TRUE, order = 0)
}
\arguments{
\item{kernel}{Kernel type. Can be a function supported on \eqn{[0, 1]}
(boundary kernel) or \eqn{[-1, 1]} (interior kernel), or else one of
\code{"triangular"} (\eqn{k(u)=(1-|u|)_{+}}), \code{"epanechnikov"}
(\eqn{k(u)=(3/4)(1-u^2)_{+}}), or \code{"uniform"} (\eqn{k(u)=
(|u|<1)/2}).}

\item{boundary}{Logical scalar, specifying whether we are at a boundary.}

\item{order}{Order of local polynomial: \code{0} means local constant,
\code{1} local linear, \code{2} local quadratic etc.}
}
\value{
Equivalent kernel function.
}
\description{
Calculates equivalent kernel for local polynomial regression.
}
\examples{
EqKern(kernel = "uniform", order = 2)
}
