% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{EqKernN}
\alias{EqKernN}
\title{Compute Equivalent kernel numerically}
\usage{
EqKernN(K, boundary = TRUE, order = 0)
}
\arguments{
\item{K}{original kernel}

\item{boundary}{Logical scalar, specifying whether we are at a boundary.}

\item{order}{Order of local polynomial: \code{0} means local constant,
\code{1} local linear, \code{2} local quadratic etc.}
}
\description{
Compute Equivalent kernel numerically
}
\keyword{internal}

