% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FindZero}
\alias{FindZero}
\title{Find interval containing zero of a function, then find the zero}
\usage{
FindZero(f, ival = 1.1, negative = TRUE)
}
\arguments{
\item{f}{function whose root we're looking for}

\item{ival}{upper endpoint of initial interval in which to search}

\item{negative}{logical: should the lower endpoint be \code{1/ival} (if the
root is guaranteed to be positive), or \code{-ival}?}
}
\description{
Given function \code{f} find \code{x0} such that \code{f(x0)==0}
}
\keyword{internal}
