% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FRDData}
\alias{FRDData}
\title{Class Constructor for \code{"FRDData"}}
\usage{
FRDData(d, cutoff)
}
\arguments{
\item{d}{list with first element corresponding to the outcome vector, second
element to the treatment vector, third element to running variable
vector, and optionally an element called \code{"sigma2"} that is a matrix
with four columns corresponding to the \code{[1, 1]}, \code{[1, 2]},
\code{[2, 1]}, and \code{[2, 2]} elements of the conditional variance
matrix of the outcome and the treatment (or an estimate of the
conditional variance matrix)}

\item{cutoff}{specifies the cutoff for the running variable}
}
\value{
An object of class \code{"FRDData"}, which is a list containing the
    following components:

    \describe{

    \item{Ym}{Matrix of outcomes and treatments for observations below
    cutoff}

    \item{Yp}{Matrix of outcomes and treatments for observations above
    cutoff}

    \item{Xm}{Running variable for observations below cutoff}

    \item{Xp}{Running variable for observations above cutoff}

    \item{sigma2m}{Matrix of conditional covariances for the outcome and the
    treatment for observations below cutoff}

    \item{sigma2p}{Matrix of conditional covariances for the outcome and the
    treatment for observations above cutoff}

    \item{orig.cutoff}{Original cutoff}

    \item{var.names}{Names of the outcome, the treatment, and the running
    variable in supplied data frame}

    }
}
\description{
Convert data to a standardized format for use with low-level functions. If
the cutoff for treatment is non-zero, shift the running variable so that the
cutoff is at zero.
}
\examples{
## Transform retirement data
d <- FRDData(rcp[, c(6, 3, 2)], cutoff=0)
## Outcome in logs
d <- FRDData(cbind(logcn=log(rcp[, 6]), rcp[, c(3, 2)]), cutoff=0)
}
\seealso{
\code{\link{RDData}} for sharp RD, and \code{\link{LPPData}} for
    inference at a point
}
