% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRD_lp.R
\name{FRDOptBW}
\alias{FRDOptBW}
\title{Optimal Bandwidth Selection in Regression Discontinuity}
\usage{
FRDOptBW(formula, data, subset, cutoff = 0, M, kern = "triangular",
  na.action, opt.criterion, bw.equal = TRUE, alpha = 0.05,
  beta = 0.8, sclass = "H", order = 1, se.initial = "EHW",
  T0 = 0)
}
\arguments{
\item{formula}{object of class \code{"formula"} (or one that can be coerced
to that class) of the form \code{outcome ~ running_variable}}

\item{data}{optional data frame, list or environment (or object coercible by
\code{as.data.frame} to a data frame) containing the outcome and running
variables in the model. If not found in \code{data}, the variables are
taken from \code{environment(formula)}, typically the environment from
which the function is called.}

\item{subset}{optional vector specifying a subset of observations to be used
in the fitting process.}

\item{cutoff}{specifies the RD cutoff in the running variable.}

\item{M}{Bound on second derivative of the conditional mean function.}

\item{kern}{specifies kernel function used in the local regression. It can
either be a string equal to \code{"triangular"} (\eqn{k(u)=(1-|u|)_{+}}),
\code{"epanechnikov"} (\eqn{k(u)=(3/4)(1-u^2)_{+}}), or \code{"uniform"}
(\eqn{k(u)= (|u|<1)/2}), or else a kernel function.}

\item{na.action}{function which indicates what should happen when the data
contain \code{NA}s. The default is set by the \code{na.action} setting of
\code{options} (usually \code{na.omit}).}

\item{opt.criterion}{Optimality criterion that bandwidth is designed to
    optimize. The options are:

   \describe{

   \item{\code{"MSE"}}{Finite-sample maximum MSE}

   \item{\code{"FLCI"}}{Length of (fixed-length) two-sided
       confidence intervals.}

   \item{\code{"OCI"}}{Given quantile of excess length of one-sided
       confidence intervals}

    }

    The methods use conditional variance given by \code{sigma2}, if supplied.
    Otherwise, for the purpose of estimating the optimal bandwidth,
    conditional variance is estimated using the method specified by
    \code{se.initial}.}

\item{bw.equal}{logical specifying whether bandwidths on either side of
cutoff should be constrained to equal to each other.}

\item{alpha}{determines confidence level, \code{1-alpha} for
constructing/optimizing confidence intervals.}

\item{beta}{Determines quantile of excess length to optimize, if bandwidth
optimizes given quantile of excess length of one-sided confidence
intervals; otherwise ignored.}

\item{sclass}{Smoothness class, either \code{"T"} for Taylor or
\code{"H"} for Hölder class.}

\item{order}{Order of local regression 1 for linear, 2 for quadratic.}

\item{se.initial}{Method for estimating initial variance for computing
    optimal bandwidth. Except for \code{"nn"}, all methods assume
    homoskedasticity on either side of cutoff (for RD), or for all data (for
    inference at a point).

\describe{

\item{"EHW"}{Based on residuals from a local linear regression using a
            triangular kernel, and and bandwidth given by a rule-of-thumb
            bandwidth (for inference at a point, see \code{\link{ROTBW.fit}}),
            or IK bandwidth (for fuzzy and sharp RD, see
            \code{\link{IKBW.fit}}). For fuzzy RD, the IK bandwidth is based
            on the reduced-form regression.}

\item{"demeaned"}{Like EHW, but instead of using the regression residuals,
            estimate \eqn{\sigma^2_i}{sigma^2_i} by subtracting the estimated
            intercept from the outcome (and not subtracting the estimated
            slope). }

\item{"Silverman"}{Use residuals from local constant regression with uniform
kernel and bandwidth selected using Silverman's rule of thumb, as in Equation
(14) in Imbens and Kalyanaraman (2012)}
\item{"SilvermanNN"}{Use Silverman's rule of thumb to pick the bandwidth, but
use nearest neighbor estimates, rather than the residuals.}

\item{"nn"}{Use nearest neighbor estimates, without assuming homoskedasticity}
}}

\item{T0}{Initial estimate of the treatment effect for calculating the
optimal bandwidth. Only relevant for Fuzzy RD.}
}
\value{
Returns an object of class \code{"RDBW"}. The function \code{print}
    can be used to obtain and print a summary of the results. An object of
    class \code{"RDBW"} is a list containing the following components:

    \describe{
    \item{\code{hp}}{bandwidth for observations above cutoff}

    \item{\code{hm}}{bandwidth for observations below cutoff, equal to \code{hp}
    unless \code{bw.equal==FALSE}}

    \item{\code{sigma2m}, \code{sigma2p}}{estimate of conditional variance
    just above and just below cutoff, \eqn{\sigma^2_+(0)} and
    \eqn{\sigma^2_{-}(0)}}

   \item{\code{f0}}{estimate of density of running variable at cutoff, if
   bandwidth computed using asymptotic method}

   \item{\code{call}}{the matched call}

   \item{\code{na.action}}{(where relevant) information on handling of missing
   data.}

   }
}
\description{
Estimate bandwidth for sharp RD based on local polynomial regression that
optimizes either maximum mean squared error, or length or quantiles of excess
length of a honest CI under second order Hölder or Taylor class.
}
\section{Note}{

\code{subset} is evaluated in the same way as variables in \code{formula},
that is first in \code{data} and then in the environment of \code{formula}.
}

\examples{
FRDOptBW(cn~retired | elig_year, data=rcp, cutoff=0, M=c(1, 0.1),
          kern="triangular", opt.criterion="FLCI")
}
\references{
{
\cite{Armstrong, Tim, and Michal Kolesár. 2018. "Optimal Inference in a Class
of Regression Models." Econometrica 86 (2): 655–83.}

\cite{Armstrong, Timothy B., and Michal Kolesár. 2019.
"Simple and Honest Confidence Intervals in Nonparametric Regression", arXiv:
1606.01200.}

\cite{Imbens, Guido, and Kalyanaraman, Karthik,
"Optimal bandwidth choice for the regression discontinuity estimator." The
Review of Economic Studies 79 (3): 933-959.}

\cite{Kolesár, Michal, and Christoph Rothe. 2018. "Inference in Regression
Discontinuity Designs with a Discrete Running Variable." American Economic
Review 108 (8): 2277–2304.}
}
}
\seealso{
\code{\link{RDHonest}}
}
