% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gss}
\alias{gss}
\title{Modified golden section for unimodal piecewise constant function}
\usage{
gss(f, xs)
}
\description{
Modified golden section for unimodal piecewise constant function
}
\keyword{internal}
