% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{KernMoment}
\alias{KernMoment}
\title{Moments of a kernel.}
\usage{
KernMoment(K, moment = 0, boundary = TRUE, type = "raw")
}
\arguments{
\item{K}{kernel function.}

\item{moment}{order \eqn{j} of moment to compute.}

\item{boundary}{Logical scalar, specifying whether we are at a boundary.}

\item{type}{Type of moment. "raw" computes \eqn{\int_X u^j k(u)}, "absolute"
computes \eqn{\int_X |u^j k(u)|}, and "raw2" computes
\eqn{\int_X u^j k(u)^2}.}
}
\value{
Integral value (a scalar).
}
\description{
Computes moments of a kernel over \eqn{X=[0, 1]} (boundary case), or
\eqn{X=[-1, 1]} (interior case),
}
\examples{
KernMoment(function(u) u<1, moment = 3, boundary = FALSE)
KernMoment(EqKern(kernel = "triangular", order = 2),
           moment = 3)
}
