% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{lee08}
\alias{lee08}
\title{Lee (2008) US House elections dataset}
\format{
A data frame with 6,558 rows and 2 variables:
\describe{
  \item{voteshare}{Vote share in next election}
  \item{margin}{Democratic margin of victory}
 }
}
\source{
Mostly Harmless Econometrics data archive,
\url{https://economics.mit.edu/faculty/angrist/data1/mhe}
}
\usage{
lee08
}
\description{
Lee (2008) US House elections dataset
}
\references{
{

\cite{David S. Lee. Randomized experiments from non-random selection in U.S.
      House elections. Journal of Econometrics, 142(2):675–697, 2008.
      \doi{10.1016/j.jeconom.2007.05.004}}

}
}
\keyword{datasets}
