% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPP_lp.R
\name{LPP_MROT.fit}
\alias{LPP_MROT.fit}
\title{Rule of thumb for choosing M}
\usage{
LPP_MROT.fit(d)
}
\arguments{
\item{d}{object of class \code{"LPPData"}}
}
\description{
Use global quartic regression to estimate a bound on the second derivative
for inference under under second order Hölder class.
}
\examples{
LPP_MROT.fit(LPPData(lee08[lee08$margin>0, ], point=0))
}
