% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPP_lp.R
\name{LPPOptBW.fit}
\alias{LPPOptBW.fit}
\title{Optimal bandwidth selection for inference at a point}
\usage{
LPPOptBW.fit(d, M, kern = "triangular", opt.criterion, alpha = 0.05,
  beta = 0.8, sclass = "H", order = 1, se.initial = "ROTEHW")
}
\arguments{
\item{d}{object of class \code{"LPPData"}}

\item{M}{Bound on second derivative of the conditional mean function.}

\item{kern}{specifies kernel function used in the local regression. It can
either be a string equal to \code{"triangular"} (\eqn{k(u)=(1-|u|)_{+}}),
\code{"epanechnikov"} (\eqn{k(u)=(3/4)(1-u^2)_{+}}), or \code{"uniform"}
(\eqn{k(u)= (|u|<1)/2}), or else a kernel function.}

\item{opt.criterion}{Optimality criterion that bandwidth is designed to
    optimize. It can either be based on exact finite-sample maximum bias and
    finite-sample estimate of variance, or asymptotic approximations to the
    bias and variance. The options are:

   \describe{

   \item{\code{"MSE"}}{Finite-sample maximum MSE}

   \item{\code{"FLCI"}}{Length of (fixed-length) two-sided
       confidence intervals.}

   \item{\code{"OCI"}}{Given quantile of excess length of one-sided
       confidence intervals}

    }

    The finite-sample methods use conditional variance given by
    \code{sigma2}, if supplied. Otherwise, for the purpose of estimating the
    optimal bandwidth, conditional variance is assumed homoscedastic, and
    estimated using a nearest neighbor estimator.}

\item{alpha}{determines confidence level, \code{1-alpha} for
constructing/optimizing confidence intervals.}

\item{beta}{Determines quantile of excess length to optimize, if bandwidth
optimizes given quantile of excess length of one-sided confidence
intervals.}

\item{sclass}{Smoothness class, either \code{"T"} for Taylor or
\code{"H"} for Hölder class.}

\item{order}{Order of local regression 1 for linear, 2 for quadratic.}

\item{se.initial}{Method for estimating initial variance for computing
    optimal bandwidth. Ignored if data already contains estimate of variance.

\describe{

\item{"ROTEHW"}{Based on residuals from a local linear regression using a
             triangular kernel and ROT bandwidth}

\item{"ROTdemeaned"}{Based on sum of squared deviations of outcome from
              estimate of intercept in local linear regression with
              triangular kenrel and ROT bandwidth}
}}
}
\value{
a list with the following elements
    \describe{
    \item{\code{h}}{Bandwidth}

    \item{\code{sigma2}}{estimate of conditional variance, from \code{d}}
   }
}
\description{
Basic computing engine called by \code{\link{LPPOptBW}} used to find
optimal bandwidth
}
\examples{
# Lee dataset
d <- LPPData(lee08[lee08$margin>0, ], point=0)
LPPOptBW.fit(d, kern = "uniform", M = 0.1, opt.criterion = "MSE")$h
}
