% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPP_lp.R
\name{LPPOptBW}
\alias{LPPOptBW}
\title{Optimal Bandwidth Selection for inference at a point}
\usage{
LPPOptBW(formula, data, subset, point = 0, M, kern = "triangular",
  na.action, opt.criterion, alpha = 0.05, beta = 0.8, sclass = "H",
  order = 1, se.initial = "ROTEHW")
}
\arguments{
\item{formula}{object of class \code{"formula"} (or one that can be coerced
to that class) of the form \code{outcome ~ independent_variable}}

\item{data}{optional data frame, list or environment (or object coercible by
\code{as.data.frame} to a data frame) containing the outcome and
independent variables in the model. If not found in \code{data}, the
variables are taken from \code{environment(formula)}, typically the
environment from which the function is called.}

\item{subset}{optional vector specifying a subset of observations to be used
in the fitting process.}

\item{point}{specifies the point \eqn{X_0} at which to do inference}

\item{M}{Bound on second derivative of the conditional mean function.}

\item{kern}{specifies kernel function used in the local regression. It can
either be a string equal to \code{"triangular"} (\eqn{k(u)=(1-|u|)_{+}}),
\code{"epanechnikov"} (\eqn{k(u)=(3/4)(1-u^2)_{+}}), or \code{"uniform"}
(\eqn{k(u)= (|u|<1)/2}), or else a kernel function.}

\item{na.action}{function which indicates what should happen when the data
contain \code{NA}s. The default is set by the \code{na.action} setting of
\code{options} (usually \code{na.omit}).}

\item{opt.criterion}{Optimality criterion that bandwidth is designed to
    optimize. It can either be based on exact finite-sample maximum bias and
    finite-sample estimate of variance, or asymptotic approximations to the
    bias and variance. The options are:

   \describe{

   \item{\code{"MSE"}}{Finite-sample maximum MSE}

   \item{\code{"FLCI"}}{Length of (fixed-length) two-sided
       confidence intervals.}

   \item{\code{"OCI"}}{Given quantile of excess length of one-sided
       confidence intervals}

    }

    The finite-sample methods use conditional variance given by
    \code{sigma2}, if supplied. Otherwise, for the purpose of estimating the
    optimal bandwidth, conditional variance is assumed homoscedastic, and
    estimated using a nearest neighbor estimator.}

\item{alpha}{determines confidence level, \code{1-alpha} for
constructing/optimizing confidence intervals.}

\item{beta}{Determines quantile of excess length to optimize, if bandwidth
optimizes given quantile of excess length of one-sided confidence
intervals.}

\item{sclass}{Smoothness class, either \code{"T"} for Taylor or
\code{"H"} for Hölder class.}

\item{order}{Order of local regression 1 for linear, 2 for quadratic.}

\item{se.initial}{Method for estimating initial variance for computing
    optimal bandwidth. Ignored if data already contains estimate of variance.

\describe{

\item{"ROTEHW"}{Based on residuals from a local linear regression using a
             triangular kernel and ROT bandwidth}

\item{"ROTdemeaned"}{Based on sum of squared deviations of outcome from
              estimate of intercept in local linear regression with
              triangular kernel and ROT bandwidth}
}}
}
\value{
Returns an object of class \code{"LPPBW"}. The function \code{print}
    can be used to obtain and print a summary of the results. An object of
    class \code{"LPPBW"} is a list containing the following components:

    \describe{
    \item{\code{h}}{Bandwidth}

    \item{\code{sigma2}}{estimate of conditional variance at a point}

   \item{\code{call}}{the matched call}

   \item{\code{na.action}}{(where relevant) information on handling of
                           missing data.}

   }
}
\description{
Estimate bandwidth based on local polynomial regression that optimizes either
maximum mean squared error, or length or quantiles of excess length of a
honest CI under second order Hölder or Taylor class.
}
\section{Note}{

\code{subset} is evaluated in the same way as variables in \code{formula},
that is first in \code{data} and then in the environment of \code{formula}.
}

\examples{

# Lee dataset
LPPOptBW(voteshare ~ margin, data = lee08, subset=margin>0,
         kern = "uniform", M = 0.1, opt.criterion = "MSE", sclass = "H")
}
\seealso{
\code{\link{LPPHonest}}
}
