% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPPfunctions.R
\name{LPPPrelimVar}
\alias{LPPPrelimVar}
\title{Compute preliminary estimate of variance}
\usage{
LPPPrelimVar(d, se.initial = "IKEHW")
}
\arguments{
\item{d}{object of class \code{"LPPData"}}

\item{se.initial}{Method for estimating initial variance for computing
    optimal bandwidth. Ignored if data already contains estimate of variance.

\describe{

\item{"ROTEHW"}{Based on residuals from a local linear regression using a
             triangular kernel and ROT bandwidth}

\item{"ROTdemeaned"}{Based on sum of squared deviations of outcome from
              estimate of intercept in local linear regression with
              triangular kernel and ROT bandwidth}
}}
}
\value{
object of class \code{"LPPData"} containing estimated variances.
}
\description{
Compute estimate of variance, which can then be used in optimal bandwidth
calculations.
}
