% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDfunctions.R
\name{LPReg}
\alias{LPReg}
\title{Local polynomial regression at a point, normalized to 0}
\usage{
LPReg(X, Y, h, K, order = 1, se.method = NULL, sigma2 = NA, J = 3)
}
\arguments{
\item{Y, X}{Outcome variable and regressor}

\item{h}{Bandwidth}

\item{K}{Kernel function}

\item{order}{Order of local regression 1 for linear, 2 for quadratic, etc.}

\item{se.method}{Vector with methods for estimating standard error of
estimate. If not specified or \code{NULL}, standard errors are not computed.
The elements of the vector can consist of the following method:
\describe{
    \item{"nn"}{Nearest neighbor method. Assumes data have been sorted so
                that \code{X[i]<X[i+1]}}
    \item{"EHW"}{Eicker-Huber-White}
    \item{"demeaned"}{Only subtract intercept}
    \item{"supplied.sigma"}{Use EHW, but instead of computing residuals,
          use estimates of \eqn{sigma^2_i} supplied by \code{sigma2}}
}}

\item{sigma2}{Optionally, supply estimates of \eqn{\sigma^{2}_{i}}}

\item{J}{Number of nearest neighbors, if "nn" is specified in \code{se.method}}
}
\description{
Calculate estimate of a function at a point and its variance given a
bandwidth using local polynomial regression of order \code{order}.
}
\keyword{internal}

