% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPRfunctions.R
\name{NPR_MROT.fit}
\alias{NPR_MROT.fit}
\title{Rule of thumb for choosing M}
\usage{
NPR_MROT.fit(d)
}
\arguments{
\item{d}{object of class \code{"RDData"}}
}
\description{
Use global quartic regression to estimate a bound on the second derivative
for inference under under second order Hölder class. For RD, use a separate
regression on either side of the cutoff
}
\examples{
NPR_MROT.fit(RDData(lee08, cutoff=0))
NPR_MROT.fit(LPPData(lee08[lee08$margin>0, ], point=0))
d <- FRDData(cbind(logcn=log(rcp[, 6 ]), rcp[, c(3, 2)]), cutoff=0)
NPR_MROT.fit(d)
}
