% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPR_lp.R
\name{NPRHonest.fit}
\alias{NPRHonest.fit}
\title{Honest inference in nonparametric regression}
\usage{
NPRHonest.fit(
  d,
  M,
  kern = "triangular",
  h,
  opt.criterion,
  bw.equal = TRUE,
  alpha = 0.05,
  beta = 0.8,
  se.method = "nn",
  J = 3,
  sclass = "H",
  order = 1,
  se.initial = "EHW",
  T0 = 0,
  T0bias = FALSE
)
}
\arguments{
\item{d}{object of class \code{"RDData"}, \code{"FRDData"}, or
\code{"LPPData"}}

\item{M}{Bound on second derivative of the conditional mean function.}

\item{kern}{specifies kernel function used in the local regression. It can
either be a string equal to \code{"triangular"} (\eqn{k(u)=(1-|u|)_{+}}),
\code{"epanechnikov"} (\eqn{k(u)=(3/4)(1-u^2)_{+}}), or \code{"uniform"}
(\eqn{k(u)= (|u|<1)/2}), or else a kernel function.}

\item{h}{bandwidth, a scalar parameter. For fuzzy or sharp RD, it can be a
named vector of length two with names \code{"p"} and \code{"m"}, in which
case the bandwidth \code{h["m"]} is used for observations below the
cutoff, and the bandwidth \code{h["p"]} is used for observations above
the cutoff. If not supplied, optimal bandwidth is computed according to
criterion given by \code{opt.criterion}.}

\item{opt.criterion}{Optimality criterion that bandwidth is designed to
    optimize. The options are:

   \describe{

   \item{\code{"MSE"}}{Finite-sample maximum MSE}

   \item{\code{"FLCI"}}{Length of (fixed-length) two-sided
       confidence intervals.}

   \item{\code{"OCI"}}{Given quantile of excess length of one-sided
       confidence intervals}

    }

    The methods use conditional variance given by \code{sigma2}, if supplied.
    Otherwise, for the purpose of estimating the optimal bandwidth,
    conditional variance is estimated using the method specified by
    \code{se.initial}.}

\item{bw.equal}{logical specifying whether bandwidths on either side of
cutoff should be constrained to equal to each other.}

\item{alpha}{determines confidence level, \code{1-alpha} for
constructing/optimizing confidence intervals.}

\item{beta}{Determines quantile of excess length to optimize, if bandwidth
optimizes given quantile of excess length of one-sided confidence
intervals; otherwise ignored.}

\item{se.method}{Vector with methods for estimating standard error of
estimate. If \code{NULL}, standard errors are not computed. The elements of
the vector can consist of the following methods:

\describe{
    \item{"nn"}{Nearest neighbor method}

    \item{"EHW"}{Eicker-Huber-White, with residuals from local regression
    (local polynomial estimators only).}

    \item{"demeaned"}{Like EHW, but instead of using the regression
        residuals, estimate \eqn{\sigma^2_i}{sigma^2_i} by subtracting the
        estimated intercept from the outcome (and not subtracting the
        estimated slope). Local polynomial estimators only.}

   \item{"plugin"}{Plug-in estimate based on asymptotic variance. Local
        polynomial estimators in sharp RD only.}

   \item{"supplied.var"}{Use conditional variance supplied by \code{sigma2} or
        \code{d} instead of computing residuals}

}}

\item{J}{Number of nearest neighbors, if "nn" is specified in
\code{se.method}.}

\item{sclass}{Smoothness class, either \code{"T"} for Taylor or
\code{"H"} for Hölder class.}

\item{order}{Order of local regression 1 for linear, 2 for quadratic.}

\item{se.initial}{Method for estimating initial variance for computing
    optimal bandwidth. Except for \code{"nn"}, all methods assume
    homoskedasticity on either side of cutoff (for RD), or for all data (for
    inference at a point).

\describe{

\item{"EHW"}{Based on residuals from a local linear regression using a
            triangular kernel, and and bandwidth given by a rule-of-thumb
            bandwidth (for inference at a point, see \code{\link{ROTBW.fit}}),
            or IK bandwidth (for fuzzy and sharp RD, see
            \code{\link{IKBW.fit}}). For fuzzy RD, the IK bandwidth is based
            on the reduced-form regression.}

\item{"demeaned"}{Like EHW, but instead of using the regression residuals,
            estimate \eqn{\sigma^2_i}{sigma^2_i} by subtracting the estimated
            intercept from the outcome (and not subtracting the estimated
            slope). }

\item{"Silverman"}{Use residuals from local constant regression with uniform
kernel and bandwidth selected using Silverman's rule of thumb, as in Equation
(14) in Imbens and Kalyanaraman (2012)}
\item{"SilvermanNN"}{Use Silverman's rule of thumb to pick the bandwidth, but
use nearest neighbor estimates, rather than the residuals.}

\item{"nn"}{Use nearest neighbor estimates, without assuming homoskedasticity}
}}

\item{T0}{Initial estimate of the treatment effect for calculating the
optimal bandwidth. Only relevant for Fuzzy RD.}

\item{T0bias}{When evaluating the maximum bias of the estimate, use the
estimate itself (if \code{T0bias==FALSE}), or use the preliminary
estimate \code{T0} (if \code{T0bias==TRUE}). Only relevant for Fuzzy RD.}
}
\value{
Returns an object of class \code{"NPRResults"}, see descriptions in
    \code{\link{RDHonest}}, \code{\link{LPPHonest}}, and
    \code{\link{FRDHonest}}.
}
\description{
Basic computing engine called by \code{\link{RDHonest}},
\code{\link{LPPHonest}}, and \code{\link{FRDHonest}} to compute honest
confidence intervals for local polynomial estimators.
}
\examples{
d <- RDData(lee08, cutoff=0)
NPRHonest.fit(d, opt.criterion="MSE", M=NPR_MROT.fit(d))
}
