% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPRfunctions.R
\name{NPRPrelimVar.fit}
\alias{NPRPrelimVar.fit}
\title{Compute preliminary estimate of variance}
\usage{
NPRPrelimVar.fit(d, se.initial = "EHW")
}
\arguments{
\item{d}{object of class \code{"RDData"}, \code{"FRDData"}, or
\code{"LPPData"}}

\item{se.initial}{Method for estimating initial variance for computing
    optimal bandwidth. Except for \code{"nn"}, all methods assume
    homoskedasticity on either side of cutoff (for RD), or for all data (for
    inference at a point).

\describe{

\item{"EHW"}{Based on residuals from a local linear regression using a
            triangular kernel, and a bandwidth given by a rule-of-thumb
            bandwidth suggested by Fan and Gijbels (1996) (for inference at a
            point), or Imbens and Kalyanaraman (2012, IK) bandwidth (for
            fuzzy and sharp RD). For fuzzy RD, the IK bandwidth is based on
            the reduced-form regression.}

\item{"demeaned"}{Like EHW, but instead of using the regression residuals,
            estimate \eqn{\sigma^2_i}{sigma^2_i} by subtracting the estimated
            intercept from the outcome (and not subtracting the estimated
            slope). }

\item{"Silverman"}{Use residuals from local constant regression with uniform
kernel and bandwidth selected using Silverman's rule of thumb, as in Equation
(14) in Imbens and Kalyanaraman (2012)}
\item{"SilvermanNN"}{Use Silverman's rule of thumb to pick the bandwidth, but
use nearest neighbor estimates, rather than the residuals.}

\item{"nn"}{Use nearest neighbor estimates, without assuming homoskedasticity}
}}
}
\value{
object of the same class as \code{d} containing estimated variances.
}
\description{
Compute estimate of variance, which can then be used in optimal bandwidth
calculations. Except for \code{se.initial="nn"}, these estimates are
unweighted.
}
