% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPRfunctions.R
\name{NPRreg.fit}
\alias{NPRreg.fit}
\title{Nonparametric Regression}
\usage{
NPRreg.fit(
  d,
  h,
  kern = "triangular",
  order = 1,
  se.method = "nn",
  no.warning = FALSE,
  J = 3
)
}
\arguments{
\item{d}{object of class \code{"LPPData"}, \code{"RDData"}, or
\code{"FRDData"}}

\item{h}{bandwidth, a scalar parameter. If not supplied, optimal bandwidth is
computed according to criterion given by \code{opt.criterion}.}

\item{kern}{specifies kernel function used in the local regression. It can
either be a string equal to \code{"triangular"} (\eqn{k(u)=(1-|u|)_{+}}),
\code{"epanechnikov"} (\eqn{k(u)=(3/4)(1-u^2)_{+}}), or \code{"uniform"}
(\eqn{k(u)= (|u|<1)/2}), or else a kernel function.}

\item{order}{Order of local regression 1 for linear, 2 for quadratic.}

\item{se.method}{Vector with methods for estimating standard error of
estimate. If \code{NULL}, standard errors are not computed. The elements of
the vector can consist of the following methods:

\describe{
    \item{"nn"}{Nearest neighbor method}

    \item{"EHW"}{Eicker-Huber-White, with residuals from local regression
    (local polynomial estimators only).}

    \item{"demeaned"}{Like EHW, but instead of using the regression
        residuals, estimate \eqn{\sigma^2_i}{sigma^2_i} by subtracting the
        estimated intercept from the outcome (and not subtracting the
        estimated slope). Local polynomial estimators only.}

   \item{"plugin"}{Plug-in estimate based on asymptotic variance. Local
        polynomial estimators in sharp RD only.}

   \item{"supplied.var"}{Use conditional variance supplied by \code{sigma2} or
        \code{d} instead of computing residuals}

}}

\item{no.warning}{Don't warn about too few observations}

\item{J}{Number of nearest neighbors, if "nn" is specified in
\code{se.method}.}
}
\value{
list with elements:

\describe{
    \item{estimate}{point estimate}
    \item{se}{Named vector of standard error estimates, as specified
              by \code{se.method}.}
    \item{w}{Implicit weight function used}

    \item{sigma2}{Estimate of \eqn{\sigma^2(X)}{sigma^2(X)} for values of
            \eqn{X} receiving positive kernel weight. By default, estimates
            are based on squared regression residuals, as used in
            \code{"EHW"}. If \code{se.method="demeaned"} or
            \code{se.method="nn"} is specified, estimates are based on that
            method, with \code{"nn"} method used if both are specified.}

     \item{eff.obs}{Number of effective observations}

}
}
\description{
Calculate fuzzy or sharp RD estimate, or estimate of a conditional mean at a
point (depending on the class of \code{d}), and its variance using local
polynomial regression of order \code{order}.
}
\examples{
NPRreg.fit(RDData(lee08, cutoff=0), h=5, order=2,
           se.method=c("nn", "plugin", "EHW"))
NPRreg.fit(LPPData(lee08[lee08$margin>=0, ], point=0), h=5, order=1)
d <- FRDData(cbind(logcn=log(rcp[, 6]), rcp[, c(3, 2)]), cutoff=0)
r <- NPRreg.fit(d, h=10, order=1)
}
