% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_RDscatter}
\alias{plot_RDscatter}
\title{Scatterplot of binned raw observations}
\usage{
plot_RDscatter(d, avg = 10, xlab = NULL, ylab = NULL, window = NULL,
  vert = TRUE)
}
\arguments{
\item{d}{RDdata object}

\item{avg}{is number of observations to average over}

\item{xlab, ylab}{x- and y-axis labels}

\item{window}{Width of a window around cutoff to which the graph should be
restricted. If not specified, full data range will be plotted}

\item{vert}{Draw a vertical line at cutoff?}
}
\description{
Scatterplot of raw observations in which each point corresponds to an binned
average.
}

