% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RD_opt.R
\name{Q}
\alias{Q}
\title{Compute inverse modulus squared divided by 4}
\usage{
Q(d, f)
}
\arguments{
\item{d}{RDData}

\item{f}{Least favorable function of class "RDLFFunction"}
}
\description{
Computes \eqn{\omega^{-1}(2b)^2/4=sum_{i} g(x_i)^2/sigma^2(x_i)},
}
\keyword{internal}
