% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RD_lp.R
\name{RD_MROT.fit}
\alias{RD_MROT.fit}
\title{Rule of thumb for choosing M}
\usage{
RD_MROT.fit(d)
}
\arguments{
\item{d}{object of class \code{"RDData"}}
}
\description{
Use global quartic regression on either side of the cutoff to estimate a
bound on the second derivative for inference under under second order Hölder
class.
}
\examples{
RD_MROT.fit(RDData(lee08, cutoff=0))
}
