% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RDData}
\alias{RDData}
\title{Class Constructor for RDData}
\usage{
RDData(d, cutoff)
}
\arguments{
\item{d}{data frame with first column corresponding to outcome variable,
second column corresponding to running variable and optionally a column
called \code{"(sigma2)"} that corresponds to the conditional variance of
the outcome (or an estimate of the conditional variance)}

\item{cutoff}{specifies the RD cutoff in the running variable.}
}
\value{
An object of class \code{"RDData"}, which is a list containing the
    following components:

    \describe{

    \item{Ym}{Outcome vector for observations below cutoff}

    \item{Yp}{Outcome vector for observationsabove cutoff}

    \item{Xm}{Running variable for observations below cutoff}

    \item{Xp}{Running variable for observationsabove cutoff}

    \item{sigma2m}{Conditional variance of outcome for observations below
    cutoff}

    \item{sigma2p}{Conditional variance of outcome for observations above
    cutoff}

    \item{orig.cutoff}{Original cutoff}

    item{var.names}{Names of outcome and running variable in supplied data frame}

    }
}
\description{
Convert data to standardized format for use with low-level functions. If the
cutoff for treatment is non-zero, shift the running variable so that cutoff
is at zero.
}
\examples{

## Transform Lee data
d <- RDData(lee08, cutoff=0)
}
