% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RD_opt.R
\name{RDEstimator}
\alias{RDEstimator}
\title{Compute optimal estimator based on solution to modulus problem, and CIs
around it}
\usage{
RDEstimator(d, f, alpha = 0.05)
}
\arguments{
\item{d}{RDData}

\item{f}{RDLFFunction}
}
\description{
\eqn{hat{L}_{delta, C}}
}
\keyword{internal}
