% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RD_opt.R
\name{RDgbC}
\alias{RDgbC}
\title{Solution to inverse modulus problem in RD under Taylor(2) class}
\usage{
RDgbC(d, b, C)
}
\arguments{
\item{d}{Object of class RDData.}

\item{b}{Jump at zero}

\item{C}{smoothness parameter for Taylor class, second derivative at zero is
bounded by \code{2*C}.}
}
\value{
Object of class \code{"RDLFFunction"}
}
\description{
Compute function \eqn{g_{b, C}(x)} that solves the inverse modulus problem
\eqn{omega^{-1}(2b,C)} in RD under second-order Taylor class with smoothness
parameter \eqn{C} by solving for parameters \eqn{d_{+}}, \eqn{d_{-}}, and
\eqn{b_{-}}
}
\keyword{internal}
