% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RD_local_linear.R
\name{RDHonest}
\alias{RDHonest}
\title{Honest inference in RD based on local polynomial estimator}
\usage{
RDHonest(formula, data, subset, cutoff = 0, M, kern = "triangular", sigma2,
  na.action, opt.criterion, bw.equal = TRUE, hp, hm = hp,
  se.method = "nn", alpha = 0.05, beta = 0.8, J = 3, sclass = "H",
  order = 1)
}
\arguments{
\item{formula}{object of class \code{"formula"} (or one that can be coerced
to that class) of the form \code{outcome ~ running_variable}}

\item{data}{optional data frame, list or environment (or object coercible by
\code{as.data.frame} to a data frame) containing the outcome and running
variables in the model. If not found in \code{data}, the variables are
taken from \code{environment(formula)}, typically the environment from
which the function is called.}

\item{subset}{optional vector specifying a subset of observations to be used
in the fitting process.}

\item{cutoff}{specifies the RD cutoff in the running variable.}

\item{M}{Bound on second derivative of the conditional mean function.}

\item{kern}{specifies kernel function used in the local regression. It
can either be a string equal to "triangular" (\eqn{k(u)=(1-|u|)_{+}}),
"epanechnikov" (\eqn{k(u)=(3/4)(1-u^2)_{+}}), or "uniform" (\eqn{k(u)=
(|u|<1)/2}), or else a function.}

\item{sigma2}{optional vector speficying the conditional variance of the
outcome variable.}

\item{na.action}{function which indicates what should happen when the data
contain \code{NA}s. The default is set by the \code{na.action} setting of
\code{options} (usually \code{na.omit}).}

\item{opt.criterion}{Optimality criterion that bandwidth is designed to
    optimize. It can either be based on exact finite-sample maximum bias and
    finite-sample estimate of variance, or asymptotic approximations to the
    bias and variance. The options are:

   \describe{

   \item{\code{"MSE"}, \code{"aMSE"}}{Finite-sample or asymtotic maximum
       MSE}

   \item{\code{"FLCI"}, \code{"aFLCI"}}{Length of (fixed-length) two-sided
       confidence intervals.}

   \item{\code{"OCI"}, \code{"aOCI"}}{Given quantile of excess length of
       one-sided confidence intervals}

    }

    The finite-sample methods use conditional variance given by
    \code{sigma2}, if supplied. Otherwise, for the purpose of estimating the
    optimal bandwidth, conditional variance is assumed homoscedastic, and
    estimated using a nearest neighbor estimator.}

\item{bw.equal}{logical specifying whether bandwidths on either side of
cutoff should be constrainted to equal to each other.}

\item{hp, hm}{bandwidth for treated (units with positive running variable),
and control (units with negative running variable) units. If \code{hm} is
not supplied, it is assumed to equal to \code{hp}. If neither bandwidth
is supplied, optimal bandwidth is computed according to criterion given
by \code{opt.criterion}.}

\item{se.method}{Vector with methods for estimating standard error of
estimate. If \code{NULL}, standard errors are not computed. The elements of
the vector can consist of the following methods:

\describe{
    \item{"nn"}{Nearest neighbor method}
    \item{"EHW"}{Eicker-Huber-White, with residuals from local regression.}
    \item{"demeaned"}{Use EHW, but instead of using residuals, estimate
        \eqn{sigma^2_i} by subtracting the estimated intercept from the
        outcome (and not subtracting the estimated slope)}
    \item{"plugin"}{Plug-in estimate based on asymptotic variance.}
    \item{"supplied.var"}{Use EHW with conditional variance supplied by
        \code{sigma2} / \code{d} instead of computing residuals}
}}

\item{alpha}{determines confidence level, \code{1-alpha} for
constructing/optimizing confidence intervals.}

\item{beta}{Determines quantile of excess length to optimize, if bandwidth
optimizes given quantile of excess length of one-sided confidence
intervals.}

\item{J}{Number of nearest neighbors, if "nn" is specified in \code{se.method}.}

\item{sclass}{Smoothness class, either \code{"T"} for Taylor or
\code{"H"} for Hölder class.}

\item{order}{Order of local regression 1 for linear, 2 for quadratic.}
}
\value{
Returns an object of class \code{"RDResults"}. The function \code{print}
    can be used to obtain and print a summary of the results. An object of
    class \code{"RDResults"} is a list containing the following components (TODO)

    \describe{
  \item{\code{estimate}}{Point estimate. This estimate is MSE-optimal if
                  \code{what="Estimation"}}

  \item{\code{lower}, \code{upper}}{Lower (upper) end-point of a one-sided CI
        based on \code{estimate}. This CI is optimal if \code{what=="OCI"}}

  \item{\code{hl}}{Half-length of a two-sided CI based on \code{estimate},
            so that the
            CI is given by \code{c(estimate-hl, estimate+hl)}. The CI is optimal
            if \code{what="FLCI"}}
  \item{\code{maxbias}}{Maximum bias of \code{estimate}}
  \item{\code{sd}}{Standard deviation of estimate}
  \item{\code{eff.obs}}{Effective number of observations used by \code{estimate}}
  \item{\code{hp}, \code{hm}}{Bandwidths used}
  \item{\code{naive}}{Coverage of naive CI that uses \code{qnorm(1-alpha/2)} as
               critical value}
}
}
\description{
Calculate estimators and one- and two-sided CIs based on local polynomial
estimator in RD under second-order Taylor or Hölder smoothness class. If
bandwidth is not supplied it is calculated to be optimal for a given
performance criterion, as specified by \code{opt.criterion} using the
function \code{\link{RDOptBW}}.
}
\section{Note}{

\code{subset} and \code{sigma2} are evaluated in the same way
as variables in \code{formula}, that is first in \code{data} and then in the
environment of \code{formula}.
}
\examples{

## TODO
}
\seealso{
\code{\link{RDOptBW}}
}

