% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RD_bme.R
\name{RDHonestBME}
\alias{RDHonestBME}
\title{BME method to compute honest CIs}
\usage{
RDHonestBME(formula, data, subset, cutoff = 0, kern = "triangular",
  na.action, bw.equal = TRUE, hp, hm = hp, alpha = 0.05, order = 1,
  regformla)
}
\description{
BME method to compute honest CIs
}
