% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RD_opt.R
\name{RDLFFunction}
\alias{RDLFFunction}
\title{Solve modulus problem in RD under second-order Taylor class}
\usage{
RDLFFunction(d, C, delta)
}
\arguments{
\item{d}{Object of class RDData.}

\item{C}{smoothness parameter for Taylor class, second derivative at zero is
bounded by \code{2*C}.}

\item{delta}{\eqn{\delta}}
}
\value{
Object of class \code{"RDLFFunction"}
}
\description{
Compute function \eqn{g_{b(\delta), C}(x)} that solves the modulus problem
\eqn{omega(\delta, C)} in RD under second-order Taylor class with smoothness
parameter \eqn{C}.
}
\keyword{internal}
