% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RD_bme.R
\name{RDlpformula}
\alias{RDlpformula}
\title{Formula for local polynomial regression}
\usage{
RDlpformula(order)
}
\arguments{
\item{order}{order of local polynomial}
}
\value{
Regression formula for local polynomial regression of order
    \code{order}
}
\description{
Formula for local polynomial regression
}
\keyword{internal}
