% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RD_lp.R
\name{RDOptBW}
\alias{RDOptBW}
\title{Optimal Bandwidth Selection in Regression Discontinuity}
\usage{
RDOptBW(formula, data, subset, cutoff = 0, M, kern = "triangular",
  na.action, opt.criterion, bw.equal = TRUE, alpha = 0.05, beta = 0.8,
  sclass = "H", order = 1, se.initial = "IKEHW")
}
\arguments{
\item{formula}{object of class \code{"formula"} (or one that can be coerced
to that class) of the form \code{outcome ~ running_variable}}

\item{data}{optional data frame, list or environment (or object coercible by
\code{as.data.frame} to a data frame) containing the outcome and running
variables in the model. If not found in \code{data}, the variables are
taken from \code{environment(formula)}, typically the environment from
which the function is called.}

\item{subset}{optional vector specifying a subset of observations to be used
in the fitting process.}

\item{cutoff}{specifies the RD cutoff in the running variable.}

\item{M}{Bound on second derivative of the conditional mean function.}

\item{kern}{specifies kernel function used in the local regression. It can
either be a string equal to \code{"triangular"} (\eqn{k(u)=(1-|u|)_{+}}),
\code{"epanechnikov"} (\eqn{k(u)=(3/4)(1-u^2)_{+}}), or \code{"uniform"}
(\eqn{k(u)= (|u|<1)/2}), or else a kernel function.}

\item{na.action}{function which indicates what should happen when the data
contain \code{NA}s. The default is set by the \code{na.action} setting of
\code{options} (usually \code{na.omit}).}

\item{opt.criterion}{Optimality criterion that bandwidth is designed to
    optimize. It can either be based on exact finite-sample maximum bias and
    finite-sample estimate of variance, or asymptotic approximations to the
    bias and variance. The options are:

   \describe{

   \item{\code{"MSE"}, \code{"aMSE"}}{Finite-sample or asymtotic maximum
       MSE}

   \item{\code{"FLCI"}, \code{"aFLCI"}}{Length of (fixed-length) two-sided
       confidence intervals.}

   \item{\code{"OCI"}, \code{"aOCI"}}{Given quantile of excess length of
       one-sided confidence intervals}

    }

    The finite-sample methods use conditional variance given by
    \code{sigma2}, if supplied. Otherwise, for the purpose of estimating the
    optimal bandwidth, conditional variance is assumed homoscedastic, and
    estimated using a nearest neighbor estimator.}

\item{bw.equal}{logical specifying whether bandwidths on either side of
cutoff should be constrainted to equal to each other.}

\item{alpha}{determines confidence level, \code{1-alpha} for
constructing/optimizing confidence intervals.}

\item{beta}{Determines quantile of excess length to optimize, if bandwidth
optimizes given quantile of excess length of one-sided confidence
intervals.}

\item{sclass}{Smoothness class, either \code{"T"} for Taylor or
\code{"H"} for Hölder class.}

\item{order}{Order of local regression 1 for linear, 2 for quadratic.}

\item{se.initial}{Method for estimating initial variance for computing
    optimal bandwidth. Ignored data already contains estimate of variance.

\describe{

\item{"IKEHW"}{Based on residuals from a local linear regression using a
              triangular kenrel and IK bandwidth}

\item{"IKdemeaned"}{Based on sum of squared deviations of outcome from
              estimate of intercept in local linear regression with
              triangular kenrel and IK bandwidth}

\item{"Silverman"}{Use residuals from local constant regression with uniform
kernel and bandwidth selected using Silverman's rule of thumb, as in Equation
(14) in IK}
\item{"SilvermanNN"}{Use nearest neighbor estimates, rather than residuals}

}}
}
\value{
Returns an object of class \code{"RDBW"}. The function \code{print}
    can be used to obtain and print a summary of the results. An object of
    class \code{"RDBW"} is a list containing the following components:

    \describe{
    \item{\code{hp}}{bandwidth for observations above cutoff}

    \item{\code{hm}}{bandwidth for observations below cutoff, equal to \code{hp}
    unless \code{bw.equal==FALSE}}

    \item{\code{sigma2m}, \code{sigma2p}}{estimate of conditional variance
    just above and just below cutoff, \eqn{\sigma^2_+(0)} and
    \eqn{\sigma^2_{-}(0)}}

   \item{\code{f0}}{estimate of density of running variable at cutoff, if
   bandwidth computed using asymptotic method}

   \item{\code{call}}{the matched call}

   \item{\code{na.action}}{(where relevant) information on handling of missing
   data.}

   }
}
\description{
Estimate bandwidth for sharp RD based on local polynomial regression that
optimizes either maximum mean squared error, or length or quantiles of excess
length of a honest CI under second order Hölder or Taylor class.
}
\section{Note}{

\code{subset} is evaluated in the same way as variables in \code{formula},
that is first in \code{data} and then in the environment of \code{formula}.
}

\examples{

## Use Lee dataset
RDOptBW(voteshare ~ margin, data = lee08, kern = "uniform",
        M = 0.1, opt.criterion = "MSE", sclass = "H")
}
\seealso{
\code{\link{RDHonest}}
}
