% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RD_opt.R
\name{RDTEfficiencyBound}
\alias{RDTEfficiencyBound}
\title{Finite-sample efficiency bounds for minimax CIs}
\usage{
RDTEfficiencyBound(object, opt.criterion = "FLCI", beta = 0.5)
}
\arguments{
\item{object}{What RDHonest returns. This has M and alpha}

\item{opt.criterion}{\code{"FLCI"} for computing efficiency of two-sided CIs,
and \code{"OCI"} for minimax one-sided CIs.}

\item{beta}{Determines quantile of excess length to optimize, if bandwidth
optimizes given quantile of excess length of one-sided confidence
intervals; otherwise ignored.}
}
\value{
Efficiency
}
\description{
Compute efficiency of minimax one-sided CIs at constant functions, or
efficiency of two-sided fixed-length CIs at constant functions under
second-order Taylor smoothness class.
}
\references{
{

\cite{Armstrong, Timothy B., and Michal Kolesár. 2018.
"Optimal Inference in a Class of Regression Models." Econometrica 86 (2):
655–83.}

}
}
