% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDfunctions.R
\name{sigmaNN}
\alias{sigmaNN}
\title{method assumes X is sorted}
\usage{
sigmaNN(X, Y, J = 3)
}
\arguments{
\item{J}{number of nearest neighbors}
}
\description{
method assumes X is sorted
}
\keyword{internal}

