% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{EqKern}
\alias{EqKern}
\title{Equivalent kernel for local linear regression.}
\usage{
EqKern(kernel = "uniform", boundary = TRUE, order = 0)
}
\arguments{
\item{kernel}{kernel type. Can be a function supported on \eqn{[-1, 1]}
(boundary kernel) or \eqn{[0,1]} (interior kernel), or else one of "uniform"
(\eqn{I(|u|<1)}), "triangular" (\eqn{(1-|u|)I(|u|<1)}), or "epanechnikov"
(k(u)=(3/4)(1-u^2)I(|u|<1)).}

\item{boundary}{Logical scalar, specifying whether we are at a boundary.}

\item{order}{Order of local polynomial: \code{0} means local constant,
\code{1} local linear, \code{2} local quadratic etc.}
}
\value{
Equivalent kernel function.
}
\description{
Calculates equivalent kernel for local polynomial regression.
}
\examples{
EqKern(kernel = "uniform", order = 2)
}
