% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RD_lp.R
\name{IKBW.fit}
\alias{IKBW.fit}
\title{Imbens and Kalyanaraman bandwidth}
\usage{
IKBW.fit(d, kern = "triangular", order = 1, verbose = FALSE)
}
\arguments{
\item{d}{object of class \code{"RDData"}}

\item{kern}{specifies kernel function used in the local regression. It
can either be a string equal to "triangular" (\eqn{k(u)=(1-|u|)_{+}}),
"epanechnikov" (\eqn{k(u)=(3/4)(1-u^2)_{+}}), or "uniform" (\eqn{k(u)=
(|u|<1)/2}), or else a function.}

\item{order}{Order of local regression 1 for linear, 2 for quadratic.}

\item{verbose}{Print details of calculation?}
}
\value{
IK bandwidth
}
\description{
Calculate bandwidth for sharp RD based on local linear regression using
method by Imbens and Kalyanaraman (2012, ReStud)
}
