% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{lee08}
\alias{lee08}
\title{Lee (2008) US House elections dataset}
\format{A data frame with 6558 rows and 2 variables:
\describe{
  \item{voteshare}{Vote share in next election}
  \item{margin}{Democratic margin of victory}
 }}
\source{
Mostly Harmless Econometrics website
}
\usage{
lee08
}
\description{
Lee (2008) US House elections dataset
}
\keyword{datasets}

