% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LPPData}
\alias{LPPData}
\title{Class Constructor for \code{"LPPData"}}
\usage{
LPPData(d, point)
}
\arguments{
\item{d}{a data frame or a list with first column corresponding to the
outcome variable, second column corresponding to the independent
variable, optionally a column called \code{"sigma2"} that corresponds to
the conditional variance of the outcome (or an estimate of the
conditional variance), and optionally a column called \code{"weights"} if
observations are aggregated by cell.}

\item{point}{specifies the point \eqn{x_0} at which to calculate the
conditional mean}
}
\value{
An object of class \code{"LPPData"}, which is a list containing the
    following components:

    \describe{

    \item{Y}{Outcome vector}

    \item{X}{Independent variable}

    \item{w}{Weights}

    \item{sigma2}{Conditional variance of the outcome}

    \item{orig.point}{Original value of \eqn{x_0}}

    \item{var.names}{Names of outcome and independent variable in supplied
    data frame}

    }
}
\description{
Convert data to standardized format for use with low-level functions. If the
point of interest \eqn{x_0} is non-zero, shift the independent variable so
that it is at zero.
}
\examples{
## Transform Lee data
d1 <- LPPData(lee08[lee08$margin>=0, ], point=0)
d2 <- LPPData(lee08, point=50)
}
\seealso{
\code{\link{FRDData}} for fuzzy RD, and \code{\link{RDData}} for
    sharp RD
}
