% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDfunctions.R
\name{LPReg}
\alias{LPReg}
\title{Local polynomial regression at a point, normalized to 0}
\usage{
LPReg(X, Y, h, K, order = 1, se.method = NULL, sigma2, J = 3)
}
\arguments{
\item{Y, X}{Outcome variable and regressor}

\item{h}{Bandwidth}

\item{K}{Kernel function}

\item{order}{Order of local regression 1 for linear, 2 for quadratic, etc.}

\item{se.method}{Vector with methods for estimating standard error of
estimate. If \code{NULL}, standard errors are not computed. The elements of
the vector can consist of the following methods:

\describe{
    \item{"nn"}{Nearest neighbor method}
    \item{"EHW"}{Eicker-Huber-White, with residuals from local regression.}
    \item{"demeaned"}{Use EHW, but instead of using residuals, estimate
        \eqn{sigma^2_i} by subtracting the estimated intercept from the
        outcome (and not subtracting the estimated slope)}
    \item{"plugin"}{Plug-in estimate based on asymptotic variance.}
    \item{"supplied.var"}{Use EHW with conditional variance supplied by
        \code{sigma2} / \code{d} instead of computing residuals}
}}

\item{sigma2}{Optionally, supply estimates of \eqn{\sigma^{2}_{i}} (for
\code{"supplied.var"} \code{se.method})}

\item{J}{Number of nearest neighbors, if "nn" is specified in \code{se.method}.}
}
\description{
Calculate estimate of a function at a point and its variance given a
bandwidth using local polynomial regression of order \code{order}.
}
\section{Note}{
 Nearest neighbor method assumes data are sorted so that
    \code{X[i] <= X[i+1]}
}

\keyword{internal}
