% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_RDscatter}
\alias{plot_RDscatter}
\title{Scatterplot of binned raw observations}
\usage{
plot_RDscatter(d, avg = 10, xlab = NULL, ylab = NULL,
  window = NULL, vert = TRUE, propdotsize = FALSE)
}
\arguments{
\item{d}{Object of class \code{"RDdata"}}

\item{avg}{Number of observations to average over. If set to \code{Inf}, then
take averages for each possible value of the running variable (convenient
when the running variable is discrete).}

\item{xlab, ylab}{x- and y-axis labels}

\item{window}{Width of a window around cutoff to which the graph should be
restricted. If not specified, full data range will be plotted}

\item{vert}{Draw a vertical line at cutoff?}

\item{propdotsize}{If \code{TRUE}, then size of points is proportional to
number of observations that the point averages over (useful when
\code{avg=Inf}). Otherwise the size of points is constant.}
}
\description{
Scatterplot of raw observations in which each point corresponds to an binned
average.
}
\examples{
plot_RDscatter(RDData(lee08, cutoff=0), avg=20)
}
