% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{rcp}
\alias{rcp}
\title{Battistin et al. (2009) retirement consumption puzzle dataset}
\format{A data frame with 30,006 rows and 6 variables:
\describe{
  \item{survey_year}{Survey year}
  \item{elig_year}{Years to/from eligibility (males)}
  \item{retired}{Retirement status (males)}
  \item{food}{Total household food expenditure}
  \item{c}{Total household consumption}
  \item{cn}{Total household nondurables expenditure}
}}
\source{
American Economic Review data archive,
\url{https://doi.org/10.1257/aer.99.5.2209}
}
\usage{
rcp
}
\description{
Battistin et al. (2009) retirement consumption puzzle dataset
}
\references{
{
\cite{Battistin, Erich, Agar Brugiavini, Enrico Rettore, and
    Guglielmo Weber. 2009. “The Retirement Consumption Puzzle: Evidence from
    a Regression Discontinuity Approach.” American Economic Review 99 (5):
    2209–26.}
}
}
\keyword{datasets}
