% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{rcp}
\alias{rcp}
\title{Battistin, Brugiavini, Rettore, and Weber (2009) retirement consumption
puzzle dataset}
\format{
A data frame with 30,006 rows and 6 variables:
\describe{
  \item{survey_year}{Survey year}
  \item{elig_year}{Years to/from eligibility (males)}
  \item{retired}{Retirement status (males)}
  \item{food}{Total household food expenditure}
  \item{c}{Total household consumption}
  \item{cn}{Total household expenditure on non-durable goods}
  \item{education}{Educational attainment (males), one of: "none",
        "elementary school", "lower secondary", "vocational studies",
        "upper secondary", "college or higher")}
  \item{family_size}{Family size}
}
}
\source{
American Economic Review data archive, \doi{10.1257/aer.99.5.2209}
}
\usage{
rcp
}
\description{
Battistin, Brugiavini, Rettore, and Weber (2009) retirement consumption
puzzle dataset
}
\references{
{

\cite{Erich Battistin, Agar Brugiavini, Enrico Rettore, and Guglielmo Weber.
     The retirement consumption puzzle: Evidence from a regression
     discontinuity approach. American Economic Review, 99(5):2209–2226, 2009.
     \doi{10.1257/aer.99.5.2209}}

}
}
\keyword{datasets}
