% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RD_opt.R
\name{RDEstimator}
\alias{RDEstimator}
\title{Compute optimal estimator based on solution to modulus problem, and CIs
around it}
\usage{
RDEstimator(d, f, alpha = 0.05, se.method = "supplied.var", J = 3)
}
\arguments{
\item{d}{RDData}

\item{f}{RDLFFunction}

\item{se.method}{Vector with methods for estimating standard error of
estimate. If \code{NULL}, standard errors are not computed. The elements of
the vector can consist of the following methods:

\describe{
    \item{"nn"}{Nearest neighbor method}

    \item{"EHW"}{Eicker-Huber-White, with residuals from local regression
    (local polynomial estimators only).}

    \item{"demeaned"}{Use EHW, but instead of using residuals, estimate
        \eqn{sigma^2_i} by subtracting the estimated intercept from the
        outcome (and not subtracting the estimated slope). Local
        polynomial estimators only.}

   \item{"plugin"}{Plug-in estimate based on asymptotic variance. Local
        polynomial estimators in RD only.}

   \item{"supplied.var"}{Use conditional variance supplied by \code{sigma2} /
        \code{d} instead of computing residuals}

}}

\item{J}{Number of nearest neighbors, if "nn" is specified in
\code{se.method}.}
}
\description{
\eqn{hat{L}_{delta, C}}
}
\keyword{internal}
