% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RD_bme.R
\name{RDHonestBME}
\alias{RDHonestBME}
\title{CIs in sharp RD with discrete regressors under bounded misspecification error
class}
\usage{
RDHonestBME(formula, data, subset, weights, cutoff = 0, na.action,
  h = Inf, alpha = 0.05, order = 0, regformula)
}
\arguments{
\item{formula}{object of class \code{"formula"} (or one that can be coerced
to that class) of the form \code{outcome ~ running_variable}}

\item{data}{optional data frame, list or environment (or object coercible by
\code{as.data.frame} to a data frame) containing the outcome and running
variables in the model. If not found in \code{data}, the variables are
taken from \code{environment(formula)}, typically the environment from
which the function is called.}

\item{subset}{optional vector specifying a subset of observations to be used
in the fitting process.}

\item{weights}{Optional vector of weights to weight the observations
(useful for aggregated data). Disregarded if optimal kernel is used.}

\item{cutoff}{specifies the RD cutoff in the running variable.}

\item{na.action}{function which indicates what should happen when the data
contain \code{NA}s. The default is set by the \code{na.action} setting of
\code{options} (usually \code{na.omit}).}

\item{h}{bandwidth, a scalar parameter. For fuzzy or sharp RD, it can be a
named vector of length two with names \code{"p"} and \code{"m"}, in which
case the bandwidth \code{h["m"]} is used for observations below the
cutoff, and the bandwidth \code{h["p"]} is used for observations above
the cutoff. If not supplied, optimal bandwidth is computed according to
criterion given by \code{opt.criterion}.}

\item{alpha}{determines confidence level, \eqn{1-\alpha}{1-alpha}}

\item{order}{Order of local regression \code{1} for linear, \code{2} for
quadratic.}

\item{regformula}{Explicitly specify regression formula to use instead of running a
local linear regression, with \code{y} and \code{x} denoting the outcome
and the running variable, and cutoff is normalized to \code{0}. Local
linear regression (\code{order = 1}) is equivalent to
\code{regformula = "y~x*I(x>0)"}. Inference is done on the \code{order+2}th
element of the design matrix}
}
\description{
Computes honest CIs for local linear regression with uniform kernel under the
bounded misspecification error class of functions, as considered in Kolesár
and Rothe (2018)
}
\details{
The parameter \code{weights} is ignored, it is only included to keep a
unified interface with \code{\link{RDHonest}}.
}
\section{Note}{

\code{subset} is evaluated in the same way as variables in \code{formula},
that is first in \code{data} and then in the environment of \code{formula}.
}

\examples{
RDHonestBME(log(cghs$earnings)~yearat14, data=cghs, h=3,
            order=1, cutoff=1947)
## Equivalent to
RDHonestBME(log(cghs$earnings)~yearat14, data=cghs, h=3,
            cutoff=1947, order=1, regformula="y~x*I(x>=0)")
}
\references{
{

\cite{Kolesár, Michal, and Christoph Rothe. 2018. "Inference in Regression
Discontinuity Designs with a Discrete Running Variable." American Economic
Review 108 (8): 2277–2304.}

}
}
