% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDfunctions.R
\name{RDLPreg}
\alias{RDLPreg}
\title{Local Polynomial Regression in Sharp RD}
\usage{
RDLPreg(d, hp, kern = "triangular", order = 1, hm = hp,
  se.method = "nn", no.warning = FALSE, J = 3)
}
\arguments{
\item{d}{object of class \code{"RDData"}}

\item{hp, hm}{bandwidth for treated (units with positive running variable),
and control (units with negative running variable) units. If \code{hm} is
not supplied, it is assumed to equal to \code{hp}. If neither bandwidth
is supplied, optimal bandwidth is computed according to criterion given
by \code{opt.criterion}.}

\item{kern}{specifies kernel function used in the local regression. It can
either be a string equal to \code{"triangular"} (\eqn{k(u)=(1-|u|)_{+}}),
\code{"epanechnikov"} (\eqn{k(u)=(3/4)(1-u^2)_{+}}), or \code{"uniform"}
(\eqn{k(u)= (|u|<1)/2}), or else a kernel function.}

\item{order}{Order of local regression 1 for linear, 2 for quadratic.}

\item{se.method}{Vector with methods for estimating standard error of
estimate. If \code{NULL}, standard errors are not computed. The elements of
the vector can consist of the following methods:

\describe{
    \item{"nn"}{Nearest neighbor method}

    \item{"EHW"}{Eicker-Huber-White, with residuals from local regression
    (local polynomial estimators only).}

    \item{"demeaned"}{Use EHW, but instead of using residuals, estimate
        \eqn{sigma^2_i} by subtracting the estimated intercept from the
        outcome (and not subtracting the estimated slope). Local
        polynomial estimators only.}

   \item{"plugin"}{Plug-in estimate based on asymptotic variance. Local
        polynomial estimators in RD only.}

   \item{"supplied.var"}{Use conditional variance supplied by \code{sigma2} /
        \code{d} instead of computing residuals}

}}

\item{no.warning}{Don't warn about too few observations}

\item{J}{Number of nearest neighbors, if "nn" is specified in
\code{se.method}.}
}
\value{
list with elements:

\describe{
    \item{estimate}{point estimate}
    \item{se}{Named vector of standard error estimates, as specified
              by \code{se.method}.}
    \item{wm,wp}{Implicit weight functions used}

    \item{sigma2m, sigma2p}{Estimates of \eqn{sigma^2(X)} for values of
            running variable below (above) cutoff and receiving positive
            kernel weight. By default, estimates are based on squared
            regression residuals, as used in \code{"EHW"}. If
            \code{"demeaned"} or \code{"nn"} is specifed, estimates are based
            on that method, with \code{"nn"} method used if both are
            specified.}

     \item{eff.obs}{Number of effective observations}

}
}
\description{
Calculate sharp RD estimate and its variance given a bandwidth using local
polynomial regression of order \code{order}.
}
