% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDfunctions.R
\name{RDPrelimVar}
\alias{RDPrelimVar}
\title{Compute preliminary estimate of variance}
\usage{
RDPrelimVar(d, se.initial = "IKEHW")
}
\arguments{
\item{d}{object of class \code{"RDData"}}

\item{se.initial}{Method for estimating initial variance for computing
    optimal bandwidth. Ignored data already contains estimate of variance.

\describe{

\item{"IKEHW"}{Based on residuals from a local linear regression using a
              triangular kenrel and IK bandwidth}

\item{"IKdemeaned"}{Based on sum of squared deviations of outcome from
              estimate of intercept in local linear regression with
              triangular kenrel and IK bandwidth}

\item{"Silverman"}{Use residuals from local constant regression with uniform
kernel and bandwidth selected using Silverman's rule of thumb, as in Equation
(14) in IK}
\item{"SilvermanNN"}{Use nearest neighbor estimates, rather than residuals}

}}
}
\value{
object of class \code{"RDData"} containing estimated variances.
}
\description{
Compute preliminary estimate of variance
}
