% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDfunctions.R
\name{RDPrelimVar}
\alias{RDPrelimVar}
\title{Compute preliminary estimate of variance}
\usage{
RDPrelimVar(d, se.initial = "IKEHW")
}
\arguments{
\item{d}{object of class \code{"RDData"}}

\item{se.initial}{Method for estimating initial variance for computing
    optimal bandwidth. Ignored if data already contains estimate of variance.

\describe{

\item{"IKEHW"}{Based on residuals from a local linear regression using a
              triangular kernel and Imbens and Kalyanaraman bandwidth}

\item{"IKdemeaned"}{Based on sum of squared deviations of outcome from
              estimate of intercept in local linear regression with
              triangular kernel and Imbens and Kalyanaraman bandwidth}

\item{"Silverman"}{Use residuals from local constant regression with uniform
kernel and bandwidth selected using Silverman's rule of thumb, as in Equation
(14) in Imbens and Kalyanaraman (2012)}
\item{"SilvermanNN"}{Use nearest neighbor estimates, rather than residuals}

\item{"NN"}{Use nearest neighbor estimates, without assuming homoscedasticity}
}}
}
\value{
object of class \code{"RDData"} containing estimated variances.
}
\description{
Compute estimate of variance, which can then be used in optimal bandwidth
calculations.
}
