% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cbound.R
\name{RDSmoothnessBound}
\alias{RDSmoothnessBound}
\title{Lower bound on smoothness constant M in RD designs}
\usage{
RDSmoothnessBound(
  d,
  s,
  separate = TRUE,
  multiple = TRUE,
  alpha = 0.05,
  sclass = "T"
)
}
\arguments{
\item{d}{object of class \code{"RDData"}}

\item{s}{Number of support points that curvature estimates should average
over}

\item{separate}{If \code{TRUE}, report estimates separately for data above
and below cutoff. If \code{FALSE}, report pooled estimates}

\item{multiple}{If \code{TRUE}, use multiple curvature estimates. If
\code{FALSE}, use a single estimate using only observations closest to
the cutoff.}

\item{alpha}{determines confidence level \code{1-alpha}.}

\item{sclass}{Smoothness class, either \code{"T"} for Taylor or \code{"H"}
for Hölder class.}
}
\value{
Returns a list with the following elements

\describe{
\item{\code{mu+,mu-}}{Lower bound of CI for observations above and below
      cutoff}

\item{\code{Z+,Z-}}{Point estimate used for lower bound}

\item{\code{sd+,sd-}}{Standard deviations of point estimates}
}
}
\description{
Estimate a lower bound on smoothness constant M and provide a lower
confidence interval.
}
\references{
{

\cite{Armstrong, Timothy B., and Michal Kolesár. 2018.
"Optimal Inference in a Class of Regression Models." Econometrica 86 (2):
655–83.}

\cite{Kolesár, Michal, and Christoph Rothe. 2018. "Inference in Regression
Discontinuity Designs with a Discrete Running Variable." American Economic
Review 108 (8): 2277–2304.}
}
}
