% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RD_opt.R
\name{RDTEfficiencyBound}
\alias{RDTEfficiencyBound}
\title{Finite-sample efficiency bounds for minimax CIs}
\usage{
RDTEfficiencyBound(d, M, opt.criterion = "FLCI", alpha = 0.05,
  beta = 0.5, se.initial = "EHW")
}
\arguments{
\item{d}{object of class \code{"RDData"}}

\item{M}{Bound on second derivative of the conditional mean function.}

\item{opt.criterion}{\code{"FLCI"} for computing efficiency of two-sided CIs,
and \code{"OCI"} for minimax one-sided CIs.}

\item{alpha}{determines confidence level, \code{1-alpha} for
constructing/optimizing confidence intervals.}

\item{beta}{Determines quantile of excess length to optimize, if bandwidth
optimizes given quantile of excess length of one-sided confidence
intervals; otherwise ignored.}

\item{se.initial}{Method for estimating initial variance for computing
    optimal bandwidth. Except for \code{"nn"}, all methods assume
    homoskedasticity on either side of cutoff (for RD), or for all data (for
    inference at a point).

\describe{

\item{"EHW"}{Based on residuals from a local linear regression using a
            triangular kernel, and and bandwidth given by a rule-of-thumb
            bandwidth (for inference at a point, see \code{\link{ROTBW.fit}}),
            or IK bandwidth (for fuzzy and sharp RD, see
            \code{\link{IKBW.fit}}). For fuzzy RD, the IK bandwidth is based
            on the reduced-form regression.}

\item{"demeaned"}{Like EHW, but instead of using the regression residuals,
            estimate \eqn{\sigma^2_i}{sigma^2_i} by subtracting the estimated
            intercept from the outcome (and not subtracting the estimated
            slope). }

\item{"Silverman"}{Use residuals from local constant regression with uniform
kernel and bandwidth selected using Silverman's rule of thumb, as in Equation
(14) in Imbens and Kalyanaraman (2012)}
\item{"SilvermanNN"}{Use Silverman's rule of thumb to pick the bandwidth, but
use nearest neighbor estimates, rather than the residuals.}

\item{"nn"}{Use nearest neighbor estimates, without assuming homoskedasticity}
}}
}
\description{
Compute efficiency of minimax one-sided CIs at constant functions, or
efficiency of two-sided fixed-length CIs at constant functions under
second-order Taylor smoothness class.
}
\references{
{

\cite{Armstrong, Tim, and Michal Kolesár. 2018. "Optimal Inference in a Class
of Regression Models." Econometrica 86 (2): 655–83.}

}
}
