% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{rebp}
\alias{rebp}
\title{Lalive (2008) Unemployment duration dataa}
\format{A data frame with 29,371 rows and 4 variables:
\describe{
  \item{age}{Age in years, at montly accuracy}
  \item{period}{Indicator for whether REBP is in place}
  \item{female}{Indicator for female}
  \item{duration}{unemployment duration in weeks}
 }}
\source{
Rafael Lalive's website
}
\usage{
rebp
}
\description{
Subset of Lalive data for individuals in the regions affected by REBP program
}
\references{
{

\cite{Lalive, R. (2008): "How Do Extended Benefits Affect Unemployment
      Duration? A Regression Discontinuity Approach." Journal of Econometrics,
      142(2): 785-806.}

}
}
\keyword{datasets}
