% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{rebp}
\alias{rebp}
\title{Austrian unemployment duration data from Lalive (2008)}
\format{
A data frame with 29,371 rows and 4 variables:
\describe{
  \item{age}{Age in years, at monthly accuracy}
  \item{period}{Indicator for whether REBP is in place}
  \item{female}{Indicator for female}
  \item{duration}{unemployment duration in weeks}
 }
}
\source{
Rafael Lalive's website,
    \url{https://sites.google.com/site/rafaellalive/}
}
\usage{
rebp
}
\description{
Subset of Lalive (2008) data for individuals in the regions affected by the
REBP program
}
\references{
{

\cite{Rafael Lalive. How do extended benefits affect unemployment duration? A
      regression discontinuity approach. Journal of Econometrics,
      142(2):785–806, February 2008. \doi{10.1016/j.jeconom.2007.05.013}}

}
}
\keyword{datasets}
